rename boundaryline app to boundaries
It may be helpful for the OSNI import
commands to share some code with BoundaryLine.

Lets give this app a more generic name
so the OSNI code can live in here too.
