--- every_election/apps/organisations/boundaryline/__init__.py
@@ -1,3 +1,0 @@
-from .boundaryline import BoundaryLine
-
-__all__ = ['BoundaryLine']
